pyproject.toml
[project]
name = "waste-finder"
version = "0.1.0"
description = "Service to download and analyze active contracts from USAspending.gov"
authors = [
    {name = "Allan Brewer", email = "allan.brewer@yahoo.com"}
]
readme = "README.md"
requires-python = "^3.11"
license = {text = "CC BY-NC 4.0"}
classifiers = [
    "License :: OSI Approved :: CC BY-NC 4.0",
]
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "python-dateutil (>=2.8.2)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "mem0ai (>=0.1.66,<0.2.0)",
    "chromadb (>=0.6.3,<0.7.0)",
    "requests-oauthlib (>=2.0.0,<3.0.0)",
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
flake8 = "^7.0.0"
isort = "^5.12.0"
ipykernel = "^6.29.5"
scikit-learn = "^1.6.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
